add dtype formatter to generate space methods of BaseModel

+++ xtools/simulation/models/model.py
@@ -24,14 +24,16 @@ class BaseModel:
         return self.observation_space.high.size
 
     @staticmethod
-    def generate_space(low, high):
+    def generate_space(low, high, dtype=np.float32):
         high = np.array(high) if type(high) is list else high
         low  = np.array(low)  if type(low)  is list else low
+        high = high.astype(dtype)
+        low  = low.astype(dtype)
         return gym.spaces.Box(high=high, low=low)
 
     @staticmethod
-    def generate_inf_space(size):
-        high = np.array([ np.inf] * size)
-        low  = np.array([-np.inf] * size)
+    def generate_inf_space(size, dtype=np.float32):
+        high = np.array([ np.inf] * size).astype(dtype)
+        low  = np.array([-np.inf] * size).astype(dtype)
         space = gym.spaces.Box(high, low)
         return high, low, space
